+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes qsf-integration project will be documented in this file.
 ## [UNRELEASED]
 ### Added
 - add userId and sessionId to the request and tracking
+- add a method to get a ObjectNode in the JsonBuilder
 
 ### Changed
 - use the document id as a fallback for tracking returned documents